*Added ability to send email notifications to topic subscribers when someone posts a reply to a topic. When a user adds a post or creates a topic, they are subscribed to that topic. Any registered user can subscribe or unsubscribe from a topic by clicking the button in the upper right of the topic.
*In Configuration area, changed SendGridUser to EmailSenderName to avoid confusion about purpose. It's the name that will appear as the sender in emails. Important: If you have a previous installation of BlazorForum, this change affects the Configuration database table and the column name will need replaced/renamed from SendGridUser to EmailSenderName.

*Updated theme for Identity Account Management area to match the main site.

*Added methods to mark user posts as deleted when the user deletes their account.

*Added methods to delete user UpDownVotes and Topic Subscriptions when the user deletes their account.

+++ BlazorForum/Areas/Identity/Pages/Account/Manage/DeletePersonalData.cshtml.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 using BlazorForum.Models;
+using BlazorForum.Domain.Interfaces;
+using BlazorForum.Domain.Helpers.Membership;
 
 namespace BlazorForum.Areas.Identity.Pages.Account.Manage
 {
@@ -14,15 +16,27 @@ namespace BlazorForum.Areas.Identity.Pages.Account.Manage
         private readonly UserManager<ApplicationUser> _userManager;
         private readonly SignInManager<ApplicationUser> _signInManager;
         private readonly ILogger<DeletePersonalDataModel> _logger;
+        private readonly IManageForumTopics _manageForumTopics;
+        private readonly IManageForumPosts _manageForumPosts;
+        private readonly IManageUpDownVotes _manageUpDownVotes;
+        private readonly IManageTopicSubscriptions _manageTopicSubscriptions;
 
         public DeletePersonalDataModel(
             UserManager<ApplicationUser> userManager,
             SignInManager<ApplicationUser> signInManager,
-            ILogger<DeletePersonalDataModel> logger)
+            ILogger<DeletePersonalDataModel> logger,
+            IManageForumTopics manageForumTopics,
+            IManageForumPosts manageForumposts,
+            IManageUpDownVotes manageUpDownVotes,
+            IManageTopicSubscriptions manageTopicSubscriptions)
         {
             _userManager = userManager;
             _signInManager = signInManager;
             _logger = logger;
+            _manageForumTopics = manageForumTopics;
+            _manageForumPosts = manageForumposts;
+            _manageUpDownVotes = manageUpDownVotes;
+            _manageTopicSubscriptions = manageTopicSubscriptions;
         }
 
         [BindProperty]
@@ -67,8 +81,17 @@ namespace BlazorForum.Areas.Identity.Pages.Account.Manage
                 }
             }
 
-            var result = await _userManager.DeleteAsync(user);
+            // Mark Topics and Posts as deleted, plus remove all UpDownVotes by user
+            var userDbFilesCleanedUp = await new UserDeleteHelpers(_manageForumTopics, _manageForumPosts, 
+                _manageUpDownVotes, _manageTopicSubscriptions)
+                .CleanupUserData(user.Id);
+
+            IdentityResult result = null;
+            if(userDbFilesCleanedUp)
+                result = await _userManager.DeleteAsync(user);
+
             var userId = await _userManager.GetUserIdAsync(user);
+
             if (!result.Succeeded)
             {
                 throw new InvalidOperationException($"Unexpected error occurred deleting user with ID '{userId}'.");
